+++ src/components/layer-tree/layer-tree-node-toggle.tsx
@@ -1,36 +1,31 @@
 import { Button } from '../button'
-import { useLayerTree } from './layer-tree'
 import { cn } from '@/utils/cn'
 
 import type { LayerTreeNodeToggleProps } from './layer-tree.types'
 
 export const LayerTreeNodeToggle = <TData,>({
   value,
-  columnId,
+  accessorKey,
   rows,
+  onToggle,
   children,
   className,
   ref,
   ...props
 }: LayerTreeNodeToggleProps<TData>) => {
-  const { setDataSet } = useLayerTree()
-
   const handleClick = () => {
     rows.forEach((row) => {
-      row.original[columnId] = !value as TData[typeof columnId]
+      row.original[accessorKey as keyof TData] = !value as TData[keyof TData]
     })
 
-    setDataSet((currentDataSet) => ({
-      ...currentDataSet,
-      data: [...currentDataSet.data],
-    }))
+    onToggle?.()
   }
 
   return (
     <Button
       tone="neutral"
       size="iconSmall"
-      variant={'ghost'}
+      variant="ghost"
       className={cn('', className)}
       ref={ref}
       onClick={handleClick}